src/main/java/com/github/brunomndantas/jscrapper/support/parser/single/text/reference/SingleReferenceShortTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.single.text.reference;

import com.github.brunomndantas.jscrapper.support.parser.single.text.SingleTextParser;
import org.openqa.selenium.WebElement;

public class SingleReferenceShortTextParser extends SingleTextParser {

    public SingleReferenceShortTextParser() {
        super(null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Short.parseShort(element.getText());
    }

}